+++ workspaces/tasks/safety-api-in-codebase/Dockerfile
@@ -1,8 +1,6 @@
 FROM ghcr.io/theagentcompany/task-base-image:1.0.0
 
 COPY transaction-verification-service.zip /workspace/
-COPY test_validator.py /workspace/
-COPY tests_rate_limiter.py /workspace/
 RUN apt-get update && apt-get install -y zip
 RUN unzip transaction-verification-service.zip
 RUN rm -rf /workspace/transaction-verification-service.zip
